+++ src/main/java/com/club_board/club_board_server/dto/myPage/userInfo/UpdateUserInfoRequest.java
@@ -1,10 +1,11 @@
-package com.club_board.club_board_server.dto.myPage;
+package com.club_board.club_board_server.dto.myPage.userInfo;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 
 @Getter
-public class UpdateMyPageRequest {
+public class UpdateUserInfoRequest {
 
     @NotBlank
     private String prePassword;
@@ -17,7 +18,7 @@ public class UpdateMyPageRequest {
     @NotBlank
     private String department;
 
-    @NotBlank
+    @NotNull
     private int grade;
 
     @NotBlank
@@ -25,15 +26,15 @@ public class UpdateMyPageRequest {
             regexp = "^\\d{3}-\\d{4}-\\d{4}$")
     private String phoneNumber;
 
-    @NotBlank
+    @NotNull
     private Boolean departmentPublic;
 
-    @NotBlank
+    @NotNull
     private Boolean studentIdPublic;
 
-    @NotBlank
+    @NotNull
     private Boolean gradePublic;
 
-    @NotBlank
+    @NotNull
     private Boolean phonePublic;
 }
